requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==24.3.0
billiard==4.2.1
celery==5.4.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
isodate==0.7.2
kombu==5.4.2
lxml==5.3.0
pillow==10.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
PyJWT==2.10.1
pyotp==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.2
redis==5.2.1
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
six==1.17.0
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
zeep==4.3.1
